requirements.txt
Django==1.6.2
Markdown==2.4
argparse==1.2.1
django-filter==0.7
djangorestframework==2.3.12
wsgiref==0.1.2
